requirements.txt
streamlit
pillow
numpy
fastembed
Qdrant_client
matplotlib
transformers
requests
openai
base64
uuid
time